+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add more bundled spinners
 - Update CLI spinners to v0.3.0 (two new spinners)
 - Rename option :use_color to just :color
+- Option to set spinner can also take frames or proc directly
 
 ### 0.1.1